pyproject.toml
[tool.poetry]
name = "blogtool"
version = "0.0.1"
description = "A tool for managing blog posts"
authors = ["Lars Kellogg-Stedman"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.12"
click = "^7.1.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
blogtool = 'blogtool.main:main'
